Proyecto 1/backend/requirements/dev.txt
flask
scikit-learn==0.22.2.post1
sqlalchemy
flask-sqlalchemy
pandas
marshmallow
flask-marshmallow
marshmallow_sqlalchemy
python-dotenv